// src/wmmt/wm.proto.d.ts
export namespace wm {

    /** Namespace protobuf. */
    namespace protobuf {
// ...
        interface ILoadCarRequest {

            /** LoadCarRequest carId */
            carId: number;

            /** LoadCarRequest eventMode */
            eventMode?: (boolean|null);

            /** LoadCarRequest eventModeSerial */
            eventModeSerial?: (string|null);
        }
// ...
    }
}